requirements.txt
requests
beautifulsoup4
flask           
readability-lxml  
